+++ src/dios-egraphs/Diospyros/llvm-tests/multiple_adds.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
-int a_in[] = {1, 2, 3, 4};
-int b_in[] = {5, 6, 7, 8};
-int c_in[] = {9, 10, 11, 12};
+float a_in[] = {1, 2, 3, 4};
+float b_in[] = {5, 6, 7, 8};
+float c_in[] = {9, 10, 11, 12};
 
 int main(int argc, char **argv) {
-    int d_out[4];
+    float d_out[4];
     d_out[0] = a_in[0] + b_in[0] + c_in[0];
     d_out[1] = a_in[1] + b_in[1] + c_in[1];
     d_out[2] = a_in[2] + b_in[2] + c_in[2];
